pyproject.toml
[project]
name = "docx-refs"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "carrascomj", email = "carrascomurielj@gmail.com" }
]
dependencies = [
    "python-docx>=1.1.2",
    "bibtexparser>=1.4.1",
    "typer>=0.12.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/docx_refs"]
